// src/app/model/amigoLogin.ts
/**
 * Digital Archive Amigo App Server API
 * An interface to a amigo app of distributed nodes for personal storage. amigo-id - hex encoded sha256 of an amigo 4096 public key amigo-key - hex encoded public key
 *
 * OpenAPI spec version: 1.0.4
 * Contact: rosborne@coredb.org
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { NodeConnection } from './nodeConnection';


export interface AmigoLogin { 
    token: string;
    account: NodeConnection;
    service: NodeConnection;
}